api/requirements.txt
fastapi
uvicorn
openai==0.28
python-multipart
pydantic
email-validator
python-dotenv
python-docx